[CoinMarketCap] Refactor more
