--- xchange-coinmarketcap/src/main/java/org/knowm/xchange/coinmarketcap/pro/v1/dto/marketdata/wrapper/CoinMarketCapCurrencyData.java
@@ -1,60 +1,0 @@
-package org.knowm.xchange.coinmarketcap.pro.v1.dto.marketdata.wrapper;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.JsonDeserializer;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import org.knowm.xchange.coinmarketcap.pro.v1.dto.marketdata.CoinMarketCapCurrencyInfo;
-
-import java.io.IOException;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.Map;
-
-@JsonDeserialize(using = CoinMarketCapCurrencyData.CoinMarketCapCurrencyDataDeserializer.class)
-public final class CoinMarketCapCurrencyData {
-
-  @JsonIgnore
-  private final Map<String, CoinMarketCapCurrencyInfo> currencyMap;
-
-  private CoinMarketCapCurrencyData(Map<String, CoinMarketCapCurrencyInfo> currencyMap) {
-    this.currencyMap = currencyMap;
-  }
-
-  public Map<String, CoinMarketCapCurrencyInfo> getCurrencyMap() {
-    return currencyMap;
-  }
-
-  @Override
-  public String toString() {
-    return "CoinMarketCapCurrencyData{" + "currencyMap=" + currencyMap + '}';
-  }
-
-  public static class CoinMarketCapCurrencyDataDeserializer
-      extends JsonDeserializer<CoinMarketCapCurrencyData> {
-
-    @Override
-    public CoinMarketCapCurrencyData deserialize(JsonParser jp, DeserializationContext ctxt)
-        throws IOException {
-      JsonNode jsonNode = jp.getCodec().readTree(jp);
-      return deserializeFromNode(jsonNode);
-    }
-
-    static CoinMarketCapCurrencyData deserializeFromNode(JsonNode jsonNode)
-        throws IOException {
-      Iterator<Map.Entry<String, JsonNode>> iterator = jsonNode.fields();
-      Map<String, CoinMarketCapCurrencyInfo> currencyMap = new HashMap<>();
-      ObjectMapper mapper = new ObjectMapper();
-      while (iterator.hasNext()) {
-        Map.Entry<String, JsonNode> entry = iterator.next();
-        CoinMarketCapCurrencyInfo currency =
-            mapper.readValue(entry.getValue().toString(), CoinMarketCapCurrencyInfo.class);
-        currencyMap.put(entry.getKey(), currency);
-      }
-      return new CoinMarketCapCurrencyData(currencyMap);
-    }
-  }
-}
